requirements.txt
beautifulsoup4
ffmpeg-python
m3u8
pycryptodome
requests
tqdm
